# Conflicts:
#	src/main/java/com/poly/petcare/app/controllers/api/CartProductControllerApi.java
#	src/main/java/com/poly/petcare/domain/entites/Order.java
#	src/main/java/com/poly/petcare/domain/repository/product/ProductRepository.java
#	src/main/java/com/poly/petcare/domain/repository/product/impl/ProductRepositoryImpl.java
#	src/main/java/com/poly/petcare/domain/services/CartProductServices.java

+++ src/main/java/com/poly/petcare/app/controllers/api/ProductControllerApi.java
@@ -61,8 +61,8 @@ public class ProductControllerApi {
     @GetMapping(value = "findByPrice")
     public ResponseEntity<?> findByPrice(@RequestParam(name = "page") Optional<Integer> page,
                                      @RequestParam(name = "limit") Optional<Integer> limit,
-                                     @RequestParam(name = "price1") Integer price1,
-                                     @RequestParam(name = "price2") Integer price2){
+                                     @RequestParam(name = "price1") BigDecimal price1,
+                                     @RequestParam(name = "price2") BigDecimal price2){
         return productService.findByPrice(page.orElse(0),limit.orElse(0),price1,price2);
     }
 }
